refactor: table-group 의 order 의존성 제거
