+++ src/main/java/kitchenpos/order/validator/OrderValidator.java
@@ -4,13 +4,16 @@ import kitchenpos.order.dao.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.validator.TableValidator;
+import kitchenpos.tableGroup.domain.TableGroup;
+import kitchenpos.tableGroup.validator.TableGroupValidator;
 import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 
 @Component
-public class OrderValidator implements TableValidator {
+public class OrderValidator implements TableValidator, TableGroupValidator {
     private final OrderRepository orderRepository;
 
     public OrderValidator(final OrderRepository orderRepository) {
@@ -28,6 +31,22 @@ public class OrderValidator implements TableValidator {
         }
         if (isPossibleChangeEmptyOrderStatus) {
             throw new IllegalArgumentException("주문 상태가 요리중 또는 식사중인 상태인 주문 테이블의 빈 자리 여부는 변경할 수 없습니다.");
+        }
+    }
+
+    @Override
+    public void possibleUngroupTableGroup(final TableGroup tableGroup) {
+        isPossibleUngroupTable(tableGroup);
+    }
+
+    private void isPossibleUngroupTable(TableGroup tableGroup) {
+        List<Long> orderTableIds = tableGroup.getOrderTables().orderTableIds();
+        boolean isImpossibleUngroupTable = orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL)
+        );
+
+        if (isImpossibleUngroupTable) {
+            throw new IllegalArgumentException("그룹 해제를 하려는 테이블 중 요리중 또는 식사중인 테이블이 존재합니다.");
         }
     }
 }
